[#162] apis에 있는 두개의 컴포넌트 mypage.ts로 묶음
