+++ src/components/mypage/MyPageStampCount.tsx
@@ -2,10 +2,10 @@
 
 import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEY } from '@/hooks/queries/query.keys';
-import { getStampCount } from '@/services/apis/stampCount';
 
 import Link from 'next/link';
 import useUserId from '@/hooks/auth/useUserId';
+import { getStampCount } from '@/services/apis/mypage';
 
 const StampCount = () => {
   const userId = useUserId();
