+++ src/services/apis/photoCount.ts
@@ -1,12 +1,12 @@
-import browserClient from '@/utils/supabase/client';
+// import browserClient from '@/utils/supabase/client';
 
-export const getPhotoCount = async (userId: string): Promise<number> => {
-  const { count, error } = await browserClient.from('album').select('*', { count: 'exact' }).eq('user_id', userId);
+// export const getPhotoCount = async (userId: string): Promise<number> => {
+//   const { count, error } = await browserClient.from('album').select('*', { count: 'exact' }).eq('user_id', userId);
 
-  if (error) {
-    console.error('사진 개수를 불러오는 중 오류:', error);
-    throw new Error('Failed to fetch photo count');
-  }
+//   if (error) {
+//     console.error('사진 개수를 불러오는 중 오류:', error);
+//     throw new Error('Failed to fetch photo count');
+//   }
 
-  return count || 0;
-};
+//   return count || 0;
+// };
