+++ src/services/apis/stampCount.ts
@@ -1,11 +1,11 @@
-import browserClient from '@/utils/supabase/client';
+// import browserClient from '@/utils/supabase/client';
 
-export const getStampCount = async (userId: string) => {
-  const { count, error } = await browserClient.from('stamp').select('*', { count: 'exact' }).eq('user_id', userId);
+// export const getStampCount = async (userId: string) => {
+//   const { count, error } = await browserClient.from('stamp').select('*', { count: 'exact' }).eq('user_id', userId);
 
-  if (error) {
-    throw new Error('stamp count를 불러오는데 실패');
-  }
+//   if (error) {
+//     throw new Error('stamp count를 불러오는데 실패');
+//   }
 
-  return count || 0;
-};
+//   return count || 0;
+// };
